Six Mens Morris Game using Std Draw
Using STD Draw which is a standard drawing library. This class provides
a basic capability for creating drawings with your programs. I used this
for a user interface and built the rest of the components including a
feature to save and load previously stored games to continue at a later
date.